refactor(server): /api/skip-events + /api/new-issue 라우트 분할 (Plan C #C9 10단계)
- src/server/routes/skip-events.ts 신규 (118줄): 3개 라우트 캡슐화
  (stats, list, delete group)
- src/server/routes/new-issue.ts 신규 (67줄): GitHub 이슈 생성 라우트
  + NewIssueRequestSchema 인라인 → 라우트 모듈로 이전
- safeError 적용
- dashboard-api.ts 7개 unused import 정리
  (resolve/join/runCli/sanitizeErrorMessage/getErrorMessage/loadTemplate/
  renderTemplate/GetSkipEventsQuerySchema)
- dashboard-api.ts 661 → 521줄 (-140)

남은 도메인: sse / auth (대시보드 thin assembler 진입)
